Solution for coding challenge
I provided some test cases that I tried myself and provided as much clarity as I could with my comments. I wanted to note that the TSV file had some inconsistencies with the company and name fields. I provided two solutions that tackle this problem differently depending on what the expected output is supposed to be.